+++ hazelcast/src/main/java/com/hazelcast/tpc/ActorMain.java
@@ -18,7 +18,7 @@ package com.hazelcast.tpc;
 
 import com.hazelcast.tpc.engine.actor.Actor;
 import com.hazelcast.tpc.engine.actor.EchoActor;
-import com.hazelcast.tpc.engine.actor.LocalActorHandle;
+import com.hazelcast.tpc.engine.actor.LocalActorRef;
 import com.hazelcast.tpc.engine.Eventloop;
 import com.hazelcast.tpc.engine.nio.NioEventloop;
 
@@ -31,7 +31,7 @@ public class ActorMain {
         Actor actor = new EchoActor();
         actor.activate(eventloop);
 
-        LocalActorHandle handle = actor.handle();
+        LocalActorRef handle = actor.handle();
         handle.send("foo");
         handle.send("bar");
     }
